Feat/add many reason in archive places by admin #769

+++ server/src/models/place/place.queries.js
@@ -1,8 +1,13 @@
+/**
+ * Ensemble des actions sur les places dans la base de données
+ * @module module:models/place/place-queries
+ */
 import mongoose from 'mongoose'
 
 import Place from './place.model'
 import '../inspecteur/inspecteur.model'
-import { appLogger } from '../../util'
+import { appLogger, techLogger } from '../../util'
+import { createArchivedPlaceFromPlace } from '../archived-place/archived-place-queries'
 
 export const PLACE_ALREADY_IN_DB_ERROR = 'PLACE_ALREADY_IN_DB_ERROR'
 
@@ -17,7 +22,31 @@ export const createPlace = async leanPlace => {
   return place.save()
 }
 
-export const deletePlace = async place => {
+/**
+ * Archiver et supprimer la place
+ * @async
+ * @function deletePlace
+ * @see [createArchivedPlaceFromPlace]{@link module:models/archived-place/archived-place-queries.createArchivedPlaceFromPlace}
+ * @param {Place~PlaceModel} place La place à supprimer
+ * @param {String[]} reasons Les raisons de la suppression
+ * @param {String} byUser L'auteur de l'action
+ * @param {Boolean} isCandilib Suppression lié à une réussite ou un echec d'un examen de candilib
+ * @return {Place~PlaceModel}
+ */
+export const deletePlace = async (place, reasons, byUser, isCandilib) => {
+  if (!place) {
+    throw new Error('No place given')
+  }
+  try {
+    await createArchivedPlaceFromPlace(place, reasons, byUser, isCandilib)
+  } catch (error) {
+    techLogger.error({
+      func: 'query-place-delete',
+      action: 'archive-place',
+      description: `Could not archive place { inspecteurId:${place.inspecteur}, centreId: ${place.centre}, date: ${place.date} }: ${error.message}`,
+      error,
+    })
+  }
   const deletedPlace = place
   await place.delete()
   return deletedPlace
